Creating Machine Learning
Unfortunately, our algorithms have a really low precision score. It may be due to the fact that I didn't change the grade_score column to a simple categorical data(fail, Pass, merit, distinction), but I left it as integers and trying to exactly predict the score of a student maybe really hard to the algorithm because of the high variation in the grade_score column.